+++ client/src/components/app/App.tsx
@@ -2,8 +2,8 @@ import React, { Component } from 'react'
 import './App.css'
 import { User } from '../types/types'
 import { getAllUsers } from '../../api/UserApi'
-import LoadingIndicator from '../Common/LoadingIndicator'
-import AppHeader from '../Common/AppHeader'
+import LoadingIndicator from '../common/LoadingIndicator'
+import AppHeader from '../common/AppHeader'
 import { Switch, Route } from 'react-router-dom'
 import UserList from '../user-list/UserList'
 import Alert from 'react-s-alert'
