+++ client/src/components/common/AppHeader.tsx
