+++ client/src/components/common/LoadingIndicator.tsx
